requirements.txt
fastapi
uvicorn
python-docx
jinja2
python-multipart
htmx
python-dotenv
google-generativeai
